SimpleIO: Allow to overwrite files that are created preliminary
Sometimes FileIO need to create preliminary file and than
flush OS cache (or do other changes to a system). This commit
allows SimpleIO to do that.

New FileIO option "overwrite" is introduced and extra code
for checking current file size to be sure that we do not
overwrite wrong file.